hashify/requirements.txt
Flask
pycryptodome

# AES-128 (16 bytes)
#aB3dE9xQ7M2Lk8Pz
# AES-192 (24 bytes)
# R9kT4M8ZQ2B7E5XW3A6C1P0Y
# AES-256 (32 bytes)
# X9Q2B7E5MZ3A6C1P0YR4K8TWDNULHSJV
